Went through a major refactoring to support the sitemap later on.
Split off document loading from the FileConverter. Markdown and
FreeMarker code is now completely separated. This is better for
understandability and maintainability in the long run.

From the set of actions that are computed a list of page changes is
computed, that is then applied to the current sitemap and saved as
the new sitemap. So at the end of the run there's now always a JSON
file in the target directory describing all pages on the site. It's not
used yet. But it will be!