More refactoring yet! Man I'm enjoying myself tonight!
- Removed the sourceRoot and targetRoot parameters from Action.perform.

The DocumentSupportFactory has become the MagistoFactory, and it has
moved one level up. Now it also creates the ActionFactory. Because of
this the source and target paths now come from a single place and
needn't be carried around all the time.

- Split up the document package.

The converter code is now in its own package. So is the loader code.