Skip index files when processing the target directory.
That is: provided that an index file wasn't provided in the source
directory...

Magisto will automatically generate index files at some point. In that
scenario there will be a target file (`index.html`) but not a
corresponding source file (e.g. `index.md`). Therefore these files must
be removed from the list of target files to process, otherwise they
will be regarded as files to be deleted. That shouldn't happen.

So the system works as follows:
- From the source paths, extract all files that are custom index files
- From the target paths, filter out all index files, unless they are
present in the set of custom index files.