Implemented a few missing tests and moved some logic around.
The builder and factory weren't properly tested yet. While writing these
test cases I found it made sense to move the check on overlapping source
and target paths into the RealMagistoBuilder class, so that it's not
possible to create a factory when they do.

Bonus: the Magisto class gets a teeny bit smaller.