Moved all Magisto application logic out of the the FileSystem interface.
The FileSystem interface now only does what its name implies: work on
the filesystem.

Part of the logic - e.g. checking the source directory and preparing
the target directory - has moved to the RealMagistoFactoryBuilder.

For the marker file (touch file) I created a separate class, called
TouchFile, to represent it.